Cleanup after port typing
