+++ crates/re_ws_comms/Cargo.toml
@@ -45,6 +45,7 @@ re_log_types = { workspace = true, features = ["serde"] }
 anyhow.workspace = true
 bincode = "1.3"
 document-features = "0.2"
+thiserror.workspace = true
 
 # Client:
 ewebsock = { version = "0.2", optional = true }
@@ -59,7 +60,6 @@ futures-util = { version = "0.3", optional = true, default-features = false, fea
   "std",
 ] }
 parking_lot = { workspace = true, optional = true }
-thiserror.workspace = true
 tokio-tungstenite = { version = "0.17.1", optional = true }
 tokio = { workspace = true, optional = true, default-features = false, features = [
   "io-std",
